refactor(routes): extract /plans into plans.js (god-file decomposition)
Route contract unchanged (284 routes), 0 /plans left inline, vitest 38 passed 268 passed . Lucy (GLM-5.2), driver-gated.
